Roorder the build stack of Dockerfiles
The old inheritance graph was:

Base ──┬──> Deep Learning ──> Logging
       └──> Spark ──────────> Logging

This forced the 'Logging' container to be built on top of two
containers, which is not well supported by Docker.

The new one is:

Base ──> Logging ──┬──> Deep Learning
                   └──> Spark

Now 'Deep Learning' and 'Spark' inherit from the 'Logging' container, meaning
we no longer have the logging container trying to build on to separate
images.